illumidesk/requirements/local.txt
-r ./base.txt

black==19.10b0
coverage==5.1
django-coverage-plugin==1.8.0
django-debug-toolbar==2.2
django-extensions==2.2.9
django-stubs==1.5.0
factory-boy==2.12.0
flake8==3.7.9
flake8-isort==3.0.0
ipdb==0.13.2
mypy==0.770
pip-tools==5.0.0
pre-commit==2.2.0
psycopg2==2.8.5 --no-binary psycopg2
pylint-celery==0.3
pylint-django==2.0.15
pytest==5.3.5
pytest-django==3.9.0
pytest-sugar==0.9.2
Sphinx==3.0.1
Werkzeug==1.0.1
